pyproject.toml
[project]
name = "dataset"
version = "0.0.0"
description = "CELINE Digital Twin API prototype"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "fastapi-keycloak>=1.1.1",
    "geoalchemy2>=0.18.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "rdflib>=7.5.0",
    "sqlalchemy>=2.0.44",
    "sqlglot>=28.1.0",
    "sqlmodel>=0.0.27",
    "typer>=0.20.0",
    "uvicorn[standard]>=0.38.0",
]

[tool.uv]  # optional, helps uv handle locking
package = true

[tool.setuptools.packages.find]
where = ["."]
include = ["celine.dataset*"]
exclude = ["tests*"]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "semantic-release>=0.1.0",
]


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
update_pyproject_toml = true 
commit_parser = "conventional"

build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "chore"]
parse_squash_commits = true
ignore_merge_commits = true

[project.scripts]
dataset-cli = "celine.dataset.cli.main:run"
